+++ lib/project_types/theme/messages/messages.rb
@@ -58,6 +58,7 @@ module Theme
 
               Options:
                 {{command:-i, --themeid=THEMEID}} Theme ID. Must be an existing theme on your store.
+                {{command:-l, --live}}            Push to your remote live theme, and update your live store.
                 {{command:-d, --development}}     Push to your remote development theme, and create it if needed.
                 {{command:-u, --unpublished}}     Create a new unpublished theme and push to it.
                 {{command:-n, --nodelete}}        Runs the push command without deleting remote files from Shopify.
@@ -188,6 +189,7 @@ module Theme
 
             Options:
               {{command:-i, --themeid=THEMEID}} The Theme ID. Must be an existing theme on your store.
+              {{command:-l, --live}}            Pull theme files from your remote live theme.
               {{command:-n, --nodelete}}        Runs the pull command without deleting local files.
 
             Run without options to select theme from a list.